delete validation error msg set

+++ app/Http/Controllers/API/User/CitiesAPIController.php
@@ -98,9 +98,14 @@ class CitiesAPIController extends Controller
      */
     public function deleteAll(Request $request)
     {
-        City::whereIn('id', $request->id)->delete();
+        if(!empty($request->id)) {
+            City::whereIn('id', $request->id)->delete();
 
-        return new DataTrueResource(true);
+            return new DataTrueResource(true);
+        }
+        else{
+            return response()->json(['error' =>config('constants.messages.delete_multiple_error')], 422);
+        }
     }
 
     /**
@@ -124,7 +129,7 @@ class CitiesAPIController extends Controller
             $path1 = $request->file('file')->store('temp');
             $path = storage_path('app') . '/' . $path1;
             $import = new CitiesImport;
-            $data = \Excel::import($import, $path);
+            $data = Excel::import($import, $path);
 
             if (count($import->getErrors()) > 0) {
                 return response()->json(['errors' => $import->getErrors()], 422);
@@ -132,7 +137,7 @@ class CitiesAPIController extends Controller
             return response()->json(['success' => true]);
         }
         else{
-            return response()->json(['errors' =>config('constants.messages.file_csv_error')], 422);
+            return response()->json(['error' =>config('constants.messages.file_csv_error')], 422);
         }
     }
 }
