+++ app/Http/Controllers/API/User/PermissionsAPIController.php
@@ -93,9 +93,14 @@ class PermissionsAPIController extends Controller
      */
     public function deleteAll(Request $request)
     {
-        Permission::whereIn('id', $request->id)->delete();
+        if(!empty($request->id)) {
+            Permission::whereIn('id', $request->id)->delete();
 
-        return new DataTrueResource(true);
+            return new DataTrueResource(true);
+        }
+        else{
+            return response()->json(['error' =>config('constants.messages.delete_multiple_error')], 422);
+        }
     }
     /**
      * This method is used set/unset permission to role
