+++ app/Http/Controllers/API/User/StatesAPIController.php
@@ -98,9 +98,14 @@ class StatesAPIController extends Controller
      */
     public function deleteAll(Request $request)
     {
-        State::whereIn('id', $request->id)->delete();
+        if(!empty($request->id)) {
+            State::whereIn('id', $request->id)->delete();
 
-        return new DataTrueResource(true);
+            return new DataTrueResource(true);
+        }
+        else{
+            return response()->json(['error' =>config('constants.messages.delete_multiple_error')], 422);
+        }
     }
     /**
      * Export States Data
